volume/layout: write MANIFEST.toml through Manifest::to_toml
generate_manifest_toml keeps its name and signature but now builds a
manifest::Manifest and returns to_toml() (the format! body moved
verbatim, so bytes are unchanged — tests/on_tape_golden.rs passes
without modification). The local ManifestUnit/ManifestSlice structs
are replaced with `pub use crate::volume::manifest::{ManifestSlice,
ManifestUnit};` so every existing layout::ManifestUnit path still
compiles. ManifestSlice::tape_position widens from i32 to i64 in the
process — the writer's i32 was the narrower of the two disagreeing
types (see the previous commit); build.rs:758's construction site
needs an i64::from(..) cast to match, and that one-line fix rides in
this commit because the widening otherwise leaves the crate not
compiling between commits.

generate_manifest_toml's manifest.units.to_vec() clones every unit and
slice per envelope written; volumes are small so this is cheap, and
the alternative (a borrowing Manifest<'a>) would change the public
type shape, which is out of scope here.

Also converts two of the three hand-typed MANIFEST.toml fixtures in
the RESTORE.sh awk tests (#131, #133) to build their scenario through
ManifestUnit/ManifestSlice and generate_manifest_toml instead, so they
track the real writer's shape. The third (#135, guarding a decoy
`name` key inside [[units.slices]]) stays hand-typed: ManifestSlice
has no such field, so the writer cannot produce that line.

+++ src/volume/build.rs
@@ -755,7 +755,7 @@ fn to_manifest_unit(unit: &BuildUnit, positions: &HashMap<i64, i32>) -> layout::
             .iter()
             .map(|sl| layout::ManifestSlice {
                 number: sl.slice_number,
-                tape_position: positions.get(&sl.slice_id).copied().unwrap_or(0),
+                tape_position: i64::from(positions.get(&sl.slice_id).copied().unwrap_or(0)),
                 size_bytes: sl.size_bytes,
                 encrypted_bytes: sl.encrypted_bytes,
                 sha256_plain: sl.sha256_plain.clone(),
